+++ src/components/Input/Input.module.scss
@@ -58,6 +58,6 @@
 
   .error_message_wrapper {
     display: block;
-    height: 14px;
+    min-height: 14px;
   }
 }
